Updating setup.py to discover data module and disambiguating from data in __init__.py
